bot/requirements.txt
tweepy
openai
python-dotenv
schedule
requests
beautifulsoup4
